url: use ES6 classes and move some functions into internals
