+++ lib/internal/url.js
@@ -1408,6 +1408,185 @@ function constructUrl(flags, protocol, username, password,
 }
 setURLConstructor(constructUrl);
 
+// Automatically escape all delimiters and unwise characters from RFC 2396.
+// Also escape single quotes in case of an XSS attack.
+// Return the escaped string.
+function autoEscapeStr(rest) {
+  var escaped = '';
+  var lastEscapedPos = 0;
+  for (var i = 0; i < rest.length; ++i) {
+    // Manual switching is faster than using a Map/Object.
+    // `escaped` contains substring up to the last escaped character.
+    switch (rest.charCodeAt(i)) {
+      case 9:   // '\t'
+        // Concat if there are ordinary characters in the middle.
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%09';
+        lastEscapedPos = i + 1;
+        break;
+      case 10:  // '\n'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%0A';
+        lastEscapedPos = i + 1;
+        break;
+      case 13:  // '\r'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%0D';
+        lastEscapedPos = i + 1;
+        break;
+      case 32:  // ' '
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%20';
+        lastEscapedPos = i + 1;
+        break;
+      case 34:  // '"'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%22';
+        lastEscapedPos = i + 1;
+        break;
+      case 39:  // '\''
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%27';
+        lastEscapedPos = i + 1;
+        break;
+      case 60:  // '<'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%3C';
+        lastEscapedPos = i + 1;
+        break;
+      case 62:  // '>'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%3E';
+        lastEscapedPos = i + 1;
+        break;
+      case 92:  // '\\'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%5C';
+        lastEscapedPos = i + 1;
+        break;
+      case 94:  // '^'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%5E';
+        lastEscapedPos = i + 1;
+        break;
+      case 96:  // '`'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%60';
+        lastEscapedPos = i + 1;
+        break;
+      case 123: // '{'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%7B';
+        lastEscapedPos = i + 1;
+        break;
+      case 124: // '|'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%7C';
+        lastEscapedPos = i + 1;
+        break;
+      case 125: // '}'
+        if (i > lastEscapedPos)
+          escaped += rest.slice(lastEscapedPos, i);
+        escaped += '%7D';
+        lastEscapedPos = i + 1;
+        break;
+    }
+  }
+  if (lastEscapedPos === 0)  // Nothing has been escaped.
+    return rest;
+
+  // There are ordinary characters at the end.
+  if (lastEscapedPos < rest.length)
+    escaped += rest.slice(lastEscapedPos);
+
+  return escaped;
+}
+
+// These characters do not need escaping:
+// ! - . _ ~
+// ' ( ) * :
+// digits
+// alpha (uppercase)
+// alpha (lowercase)
+const noEscapeAuth = [
+  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 0x00 - 0x0F
+  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 0x10 - 0x1F
+  0, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 1, 1, 0, // 0x20 - 0x2F
+  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, // 0x30 - 0x3F
+  0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 0x40 - 0x4F
+  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, // 0x50 - 0x5F
+  0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 0x60 - 0x6F
+  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0  // 0x70 - 0x7F
+];
+
+function urlEncodeAuth(str) {
+  // faster encodeURIComponent alternative for encoding auth uri components
+  var out = '';
+  var lastPos = 0;
+  for (var i = 0; i < str.length; ++i) {
+    var c = str.charCodeAt(i);
+
+    // ASCII
+    if (c < 0x80) {
+      if (noEscapeAuth[c] === 1)
+        continue;
+      if (lastPos < i)
+        out += str.slice(lastPos, i);
+      lastPos = i + 1;
+      out += hexTable[c];
+      continue;
+    }
+
+    if (lastPos < i)
+      out += str.slice(lastPos, i);
+
+    // Multi-byte characters ...
+    if (c < 0x800) {
+      lastPos = i + 1;
+      out += hexTable[0xC0 | (c >> 6)] + hexTable[0x80 | (c & 0x3F)];
+      continue;
+    }
+    if (c < 0xD800 || c >= 0xE000) {
+      lastPos = i + 1;
+      out += hexTable[0xE0 | (c >> 12)] +
+        hexTable[0x80 | ((c >> 6) & 0x3F)] +
+        hexTable[0x80 | (c & 0x3F)];
+      continue;
+    }
+    // Surrogate pair
+    ++i;
+    var c2;
+    if (i < str.length)
+      c2 = str.charCodeAt(i) & 0x3FF;
+    else
+      c2 = 0;
+    lastPos = i + 1;
+    c = 0x10000 + (((c & 0x3FF) << 10) | c2);
+    out += hexTable[0xF0 | (c >> 18)] +
+      hexTable[0x80 | ((c >> 12) & 0x3F)] +
+      hexTable[0x80 | ((c >> 6) & 0x3F)] +
+      hexTable[0x80 | (c & 0x3F)];
+  }
+  if (lastPos === 0)
+    return str;
+  if (lastPos < str.length)
+    return out + str.slice(lastPos);
+  return out;
+}
+
 module.exports = {
   toUSVString,
   getPathFromURL,
@@ -1417,6 +1596,8 @@ module.exports = {
   domainToASCII,
   domainToUnicode,
   urlToOptions,
+  autoEscapeStr,
+  encodeAuth: urlEncodeAuth,
   formatSymbol: kFormat,
   searchParamsSymbol: searchParams
 };
